Rip out old type checker, replace with bidir
