+++ executables/Main.hs
@@ -22,7 +22,6 @@ import System.IO (hPutStrLn, stderr)
 import Text.Megaparsec
 
 import Amy.ANF as ANF
-import Amy.Bidirectional as B
 import Amy.Codegen
 import Amy.Core as C
 import Amy.Errors
@@ -52,9 +51,9 @@ process filePath DumpFlags{..} input = do
     renamed <- liftEither $ rename parsed
 
     -- Type checking
-    typeChecked <- liftEither $ first (:[]) $ B.inferModule renamed
+    typeChecked <- liftEither $ first (:[]) $ T.inferModule renamed
     when dfDumpTypeChecked $
-      lift $ putStrLn "\nType Checked:" >> print (B.prettyModule typeChecked)
+      lift $ putStrLn "\nType Checked:" >> print (T.prettyModule typeChecked)
 
     -- Desugar to Core
     let core = desugarModule typeChecked
