--- library/Amy/Bidirectional.hs
@@ -1,7 +1,0 @@
-module Amy.Bidirectional
-  ( module X
-  ) where
-
-import Amy.Bidirectional.AST as X
-import Amy.Bidirectional.TypeCheck as X
-import Amy.Bidirectional.Pretty as X
