--- library/Amy/Bidirectional/Pretty.hs
@@ -1,102 +1,0 @@
-{-# LANGUAGE OverloadedStrings #-}
-
-module Amy.Bidirectional.Pretty
-  ( prettyModule
-  , prettyExpr
-
-  , prettyType
-  ) where
-
-import Data.Foldable (toList)
-import qualified Data.Map.Strict as Map
-
-import Amy.Literal
-import Amy.Pretty
-import Amy.Bidirectional.AST
-
-prettyType :: Type -> Doc ann
-prettyType (TyFun ty1 ty2) = parensIf (isTyFun ty1) (prettyType ty1) <+> "->" <+> prettyType ty2
-prettyType (TyCon con) = prettyTyConName con
-prettyType (TyVar var) = prettyTyVarName var
-prettyType (TyExistVar var) = prettyTyExistVarName var
-prettyType (TyRecord rows mVar) = prettyTyRecord (uncurry prettyTyRow <$> Map.toList rows) (prettyType <$> mVar)
-prettyType (TyApp f arg) = prettyType f <+> parensIf (isTyApp arg) (prettyType arg)
-prettyType (TyForall vars ty) = "forall" <+> hcat (punctuate space $ prettyTyVarName <$> toList vars) <> "." <+> prettyType ty
-
-prettyTyRow :: RowLabel -> Type -> Doc ann
-prettyTyRow label ty = prettyRowLabel label <+> "::" <+> prettyType ty
-
-isTyApp :: Type -> Bool
-isTyApp TyApp{} = True
-isTyApp TyFun{} = True
-isTyApp _ = False
-
-isTyFun :: Type -> Bool
-isTyFun TyFun{} = True
-isTyFun _ = False
-
-prettyModule :: Module -> Doc ann
-prettyModule (Module bindings externs typeDeclarations) =
-  vcatTwoHardLines
-  $ (prettyExtern' <$> externs)
-  ++ (prettyTypeDeclaration' <$> typeDeclarations)
-  ++ (prettyBinding' <$> bindings)
-
-prettyExtern' :: Extern -> Doc ann
-prettyExtern' (Extern name ty) =
-  prettyExtern (prettyIdent name) (prettyType ty)
-
-prettyTypeDeclaration' :: TypeDeclaration -> Doc ann
-prettyTypeDeclaration' (TypeDeclaration tyName cons) =
-   prettyTypeDeclaration (prettyTyConDefinition tyName) (prettyConstructor <$> cons)
- where
-  prettyConstructor (DataConDefinition conName mArg) =
-    prettyDataConstructor (prettyDataConName conName) (prettyType <$> mArg)
-
-prettyTyConDefinition :: TyConDefinition -> Doc ann
-prettyTyConDefinition (TyConDefinition name args) = prettyTyConName name <> args'
- where
-  args' = if null args then mempty else space <> sep (prettyTyVarName <$> args)
-
-prettyBinding' :: Binding -> Doc ann
-prettyBinding' (Binding ident ty args _ body) =
-  prettyBindingType' ident ty <>
-  hardline <>
-  prettyBinding (prettyIdent ident) (prettyIdent . typedValue <$> args) (prettyExpr body)
-
-prettyBindingType' :: IdentName -> Type -> Doc ann
-prettyBindingType' ident scheme = prettyBindingType (prettyIdent ident) (prettyType scheme)
-
-prettyExpr :: Expr -> Doc ann
-prettyExpr (ELit lit) = pretty $ showLiteral lit
-prettyExpr (ERecord rows) = bracketed $ uncurry prettyRow <$> Map.toList (typedValue <$> rows)
-prettyExpr (ERecordSelect expr field _) = prettyExpr expr <> "." <> prettyRowLabel field
-prettyExpr (EVar var) = prettyVar var
-prettyExpr (EIf (If pred' then' else')) =
-  prettyIf (prettyExpr pred') (prettyExpr then') (prettyExpr else')
-prettyExpr (ECase (Case scrutinee matches)) =
-  prettyCase (prettyExpr scrutinee) Nothing (toList $ mkMatch <$> matches)
- where
-  mkMatch (Match pat body) = (prettyPattern pat, prettyExpr body)
-prettyExpr (ELet (Let bindings body)) =
-  prettyLet (prettyBinding' <$> bindings) (prettyExpr body)
-prettyExpr (EApp (App f arg _)) = prettyExpr f <+> prettyExpr arg
-prettyExpr (EParens expr) = parens $ prettyExpr expr
-
-prettyRow :: RowLabel -> Expr -> Doc ann
-prettyRow label expr = prettyRowLabel label <+> "=" <+> prettyExpr expr
-
-prettyVar :: Var -> Doc ann
-prettyVar (VVal (Typed _ var)) = prettyIdent var
-prettyVar (VCons (Typed _ con)) = prettyDataConName con
-
-prettyPattern :: Pattern -> Doc ann
-prettyPattern (PLit lit) = pretty $ showLiteral lit
-prettyPattern (PVar (Typed _ var)) = prettyIdent var
-prettyPattern (PParens pat) = parens (prettyPattern pat)
-prettyPattern (PCons (PatCons con mArg _)) =
-  prettyDataConName con <> maybe mempty prettyArg mArg
- where
-  prettyArg = (space <>) . prettyArg'
-  prettyArg' arg@PCons{} = parens (prettyPattern arg)
-  prettyArg' arg = prettyPattern arg
