+++ library/Amy/TypeCheck/KindInference.hs
@@ -2,10 +2,12 @@
 
 module Amy.TypeCheck.KindInference
   ( inferTypeDeclarationKind
+  , checkTypeKind
   , inferTypeKind
   ) where
 
 import Control.Monad.Except
+import Data.Foldable (traverse_)
 import Data.Map.Strict (Map)
 import qualified Data.Map.Strict as Map
 import Data.Maybe (fromMaybe, mapMaybe, maybeToList)
@@ -25,7 +27,7 @@ import Amy.TypeCheck.Monad
 -- TODO: Infer the kinds of mutually recursive type declaration groups at the
 -- same time instead of one by one.
 
-inferTypeDeclarationKind :: TypeDeclaration -> Inference Kind
+inferTypeDeclarationKind :: TypeDeclaration -> Checker Kind
 inferTypeDeclarationKind (TypeDeclaration (TyConDefinition tyCon tyArgs) constructors) =
   withNewLexicalScope $ do
     -- Generate unknown kind variables for the type constructor and all type
@@ -45,17 +47,23 @@ inferTypeDeclarationKind (TypeDeclaration (TyConDefinition tyCon tyArgs) constru
     let kind = fromMaybe (error "Lost the input kind") $ Map.lookup tyConKindVar subst
     pure $ starIfUnknown kind
 
-inferTypeKind :: Type -> Inference Kind
+checkTypeKind :: Type -> Checker ()
+checkTypeKind ty = do
+  kind <- inferTypeKind ty
+  when (kind /= KStar) $
+    error $ "Somehow kind unification passed but we don't have KStar " ++ show (ty, kind)
+
+inferTypeKind :: Type -> Checker Kind
 inferTypeKind ty = do
   (kind, cons) <- inferTypeKind' ty
   subst <- solver (emptySubst, cons)
   pure $ starIfUnknown $ substituteKind subst kind
 
-inferTypeKind' :: Type -> Inference (Kind, [Constraint])
+inferTypeKind' :: Type -> Checker (Kind, [Constraint])
 inferTypeKind' (TyCon name) = do
   kind <- lookupTyConKind name
   pure (kind, [])
-inferTypeKind' (TyVar (TyVarInfo name _)) = do
+inferTypeKind' (TyVar name) = do
   kind <- lookupTyVarKind name
   pure (kind, [])
 inferTypeKind' (TyApp t1 t2) = do
@@ -75,15 +83,19 @@ inferTypeKind' (TyFun t1 t2) = do
       , Constraint (kind, KStar)
       ]
   pure (kind, cons1 ++ cons2 ++ constraints)
-inferTypeKind' (TyRecord fields mVar) = do
+inferTypeKind' (TyRecord fields mTail) = do
   fieldCons <- for (Map.elems fields) $ \ty -> do
     (fieldKind, fieldCons) <- inferTypeKind' ty
     pure $ fieldCons ++ [Constraint (fieldKind, KStar)]
-  varCons <- for (maybeToList mVar) $ \(TyVarInfo var _) -> do
-    varKind <- lookupTyVarKind var
-    pure $ Constraint (varKind, KRow)
+  varCons <- for (maybeToList mTail) $ \tail' -> do
+    (kind, cons) <- inferTypeKind' tail'
+    pure $ cons ++ [Constraint (kind, KRow)]
   kind <- KUnknown <$> freshId
-  pure (kind, concat fieldCons ++ varCons ++ [Constraint (kind, KStar)])
+  pure (kind, concat fieldCons ++ concat varCons ++ [Constraint (kind, KStar)])
+inferTypeKind' (TyForall vars ty) = withNewLexicalScope $ do
+  traverse_ addUnknownTyVarKindToScope vars
+  inferTypeKind' ty
+inferTypeKind' v@(TyExistVar _) = error $ "Found existential variable in kind inference " ++ show v
 
 -- | If we have any unknown kinds left, just call them KStar.
 starIfUnknown :: Kind -> Kind
@@ -99,7 +111,7 @@ starIfUnknown (KFun k1 k2) = KFun (starIfUnknown k1) (starIfUnknown k2)
 newtype Constraint = Constraint (Kind, Kind)
   deriving (Show, Eq)
 
-solver :: (Subst, [Constraint]) -> Inference Subst
+solver :: (Subst, [Constraint]) -> Checker Subst
 solver (su, cs) =
   case cs of
     [] -> return su
@@ -107,7 +119,7 @@ solver (su, cs) =
       su1 <- unifyKinds t1 t2
       solver (su1 `composeSubst` su, substituteConstraint su1 <$> cs0)
 
-unifyKinds :: Kind -> Kind -> Inference Subst
+unifyKinds :: Kind -> Kind -> Checker Subst
 unifyKinds k1 k2 | k1 == k2 = pure emptySubst
 unifyKinds (KUnknown i) k = i `bind` k
 unifyKinds k (KUnknown i) = i `bind` k
@@ -117,7 +129,7 @@ unifyKinds (KFun k1 k2) (KFun k3 k4) = do
   pure (su2 `composeSubst` su1)
 unifyKinds k1 k2 = throwError $ KindUnificationFail k1 k2
 
-bind :: Int -> Kind -> Inference Subst
+bind :: Int -> Kind -> Checker Subst
 bind i k
   | k == KUnknown i = pure emptySubst
   | occursCheck i k = throwError $ InfiniteKind i k
